App/src/requirements.txt
Flask==3.0.0
flask-swagger-ui==4.11.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
azure-identity==1.15.0